Renaming file name and its references.
